src/chatdi/requirements.txt
Django==5.2.3
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
drf-spectacular==0.28.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
pyTelegramBotAPI==4.27.0
requests==2.32.4